+++ src/main/java/de/alpharogroup/spring/batch/factory/SpringBatchObjectFactory.java
@@ -42,31 +42,42 @@ import org.springframework.core.io.Resource;
 
 import de.alpharogroup.reflection.ReflectionExtensions;
 import de.alpharogroup.spring.batch.mapper.CustomBeanWrapperFieldSetMapper;
+import lombok.NonNull;
 import lombok.experimental.UtilityClass;
 
 @UtilityClass
 public class SpringBatchObjectFactory
 {
 
-	private static final String READER_SUFFIX = "Reader";
+	public static final String READER_SUFFIX = "Reader";
+	public static final String WRITER_SUFFIX = "Writer";
 
-	public static <T> JdbcBatchItemWriter<T> newJdbcBatchItemWriter(DataSource dataSource,
-		String sql)
+	public static <T> JdbcBatchItemWriter<T> newJdbcBatchItemWriter(
+		final @NonNull DataSource dataSource, final @NonNull String sql)
 	{
 		return new JdbcBatchItemWriterBuilder<T>()
 			.itemSqlParameterSourceProvider(new BeanPropertyItemSqlParameterSourceProvider<>())
 			.sql(sql).dataSource(dataSource).build();
 	}
 
-	public static <T> FlatFileItemReader<T> newCsvFileItemReader(Resource resource,
-		Class<T> typeClass, String delimiter, int linesToSkip)
+	public static <T> FlatFileItemReader<T> newCsvFileItemReader(final @NonNull Resource resource,
+		final @NonNull Class<T> typeClass, final @NonNull String delimiter, final int linesToSkip)
 	{
-		return newCsvFileItemReader(resource, typeClass, new CustomBeanWrapperFieldSetMapper<>(
-			typeClass, DateTimeFormatter.ofPattern("dd-MM-yyyy")), delimiter, linesToSkip);
+		return newCsvFileItemReader(resource, typeClass, DateTimeFormatter.ofPattern("dd-MM-yyyy"),
+			delimiter, linesToSkip);
 	}
 
-	public static <T> FlatFileItemReader<T> newCsvFileItemReader(Resource resource,
-		Class<T> typeClass, FieldSetMapper<T> fieldSetMapper, String delimiter, int linesToSkip)
+	public static <T> FlatFileItemReader<T> newCsvFileItemReader(final @NonNull Resource resource,
+		final @NonNull Class<T> typeClass, final @NonNull DateTimeFormatter formatter,
+		final @NonNull String delimiter, final int linesToSkip)
+	{
+		return newCsvFileItemReader(resource, typeClass,
+			new CustomBeanWrapperFieldSetMapper<>(typeClass, formatter), delimiter, linesToSkip);
+	}
+
+	public static <T> FlatFileItemReader<T> newCsvFileItemReader(final @NonNull Resource resource,
+		final @NonNull Class<T> typeClass, final @NonNull FieldSetMapper<T> fieldSetMapper,
+		final @NonNull String delimiter, final int linesToSkip)
 	{
 		DefaultLineMapper<T> lineMapper = new DefaultLineMapper<>();
 		String[] fieldNames = ReflectionExtensions.getDeclaredFieldNames(typeClass);
@@ -78,11 +89,18 @@ public class SpringBatchObjectFactory
 			.resource(resource).lineMapper(lineMapper).linesToSkip(linesToSkip).build();
 	}
 
-	public static <T> JpaItemWriter<T> newJpaItemWriter(EntityManagerFactory entityManagerFactory)
+	public static <T> JpaItemWriter<T> newJpaItemWriter(
+		final @NonNull EntityManagerFactory entityManagerFactory)
 	{
 		JpaItemWriter<T> writer = new JpaItemWriter<>();
 		writer.setEntityManagerFactory(entityManagerFactory);
 		return writer;
+	}
+
+	public static <T> CustomBeanWrapperFieldSetMapper<T> newCustomBeanWrapperFieldSetMapper(
+		final @NonNull Class<? extends T> typeClass, final @NonNull DateTimeFormatter formatter)
+	{
+		return new CustomBeanWrapperFieldSetMapper<>(typeClass, formatter);
 	}
 
 }